- rearrange TerrainType - new Player structure - fix snippet `@type` - try fix debugger stop at first line of file while import
