+++ .babelrc
@@ -13,27 +13,23 @@
       {
         "loose": true
       }
-    ]
+    ],
+    // Maybe change to `@babel/plugin-proposal-export-default-from`
+    // see: https://babeljs.io/docs/en/babel-plugin-syntax-export-default-from#syntax-only
+    "@babel/plugin-syntax-export-default-from",
+    "@babel/plugin-syntax-dynamic-import"
   ]
 }
 
-  // "plugins": [
-  //   [
-  //     "@babel/plugin-proposal-class-properties",
-  //     {
-  //       "loose": true
-  //     }
-  //   ],
-  //   "@babel/plugin-proposal-optional-chaining",
-  //   [
-  //     "@babel/plugin-proposal-decorators",
-  //     {
-  //       "legacy": true
-  //     }
-  //   ],
-  //   "@babel/plugin-proposal-function-bind",
-  //   "@babel/plugin-syntax-export-default-from",
-  //   "@babel/plugin-syntax-dynamic-import",
-  //   "@babel/plugin-syntax-flow",
-  //   "@babel/plugin-transform-runtime"
-  // ]
+// "plugins": [
+//   "@babel/plugin-proposal-optional-chaining",
+//   [
+//     "@babel/plugin-proposal-decorators",
+//     {
+//       "legacy": true
+//     }
+//   ],
+//   "@babel/plugin-proposal-function-bind",
+//   "@babel/plugin-syntax-flow",
+//   "@babel/plugin-transform-runtime"
+// ]
