+++ .vscode/launch.json
@@ -5,7 +5,22 @@
   "version": "0.2.0",
   "configurations": [
     {
-      "name": "nodemon",
+      "name": "Launch",
+      "type": "node",
+      "request": "launch",
+      "program": "${workspaceFolder}/dist/server.bundle.js",
+      "sourceMaps": true,
+      "smartStep": true,
+      "stopOnEntry": false,
+      "outFiles": [
+        "${workspaceRoot}/dist/*.js"
+      ],
+      "skipFiles": [
+        "<node_internals>/**"
+      ]
+    },
+    {
+      "name": "Launch with nodemon",
       "type": "node",
       "request": "launch",
       "runtimeExecutable": "nodemon",
