+++ src/common/core/constant/TerrainType.js
@@ -1,12 +1,12 @@
 import Enum from "../../util/Enum";
 
 let TerrainType = {
-  resource: 0,
-  plain: 1,
-  forest: 2,
-  mountain: 3,
-  ocean: 4,
-  boss: 5
+  resource: 1,
+  plain: 2,
+  forest: 3,
+  mountain: 4,
+  ocean: 5,
+  boss: 6
 };
 
 TerrainType = new Enum(TerrainType);
