+++ src/common/core/data/player/PlayerAbility.js
@@ -1,0 +1,6 @@
+export default class PlayerAbility {
+  constructor() {
+    this.canHarvestStone = false;
+    this.canHarvestDiamond = false;
+  }
+}
