+++ src/common/core/data/player/PlayerResource.js
@@ -1,0 +1,8 @@
+export default class PlayerResource {
+  constructor() {
+    this.wood = 0;
+    this.apple = 0;
+    this.stone = 0;
+    this.diamond = 0;
+  }
+}
